# public/docs/api-docs.json
{
  "swagger": "2.0",
  "info": {
    "description": "Penjualanapp API",
    "version": "1.0.0",
    "title": "Penjualanapp API",
    "contact": {
      "name": "PenjualanAPP",
      "url": "https://penjualanapp-api.herokuapp.com"
    },
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "basePath": "/api/v1",
  "schemes": [
    "http",
    "https"
  ],
  "produces": [
    "application/json"
  ],
  "consumes": [
    "application/json"
  ],
  "securityDefinitions": {
    "Bearer": {
      "description": "For accessing the protected API resources, you must have received a a valid JWT token after registering or logging in. This JWT token must then be used for all protected resources by passing it in via the 'Authorization' header.\n\nA JWT token is generated by the API by either registering via /users or logging in via /users/login.\n\nThe following format must be in the 'Authorization' header :\n\n    Token:  Bearer <token>\n    \n",
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "paths": {
...
    "/customers": {
      "get": {
        "summary": "Get all customer",
        "description": "Get Customers",
        "tags": [
          "Customer"
        ],
        "security": [
          {
            "Bearer": []
          }
        ],
        "operationId": "GetCustomer",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/CustomerResponse"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "422": {
            "description": "Unexpected error",
            "schema": {
              "$ref": "#/definitions/GenericErrorModel"
            }
          }
        }
      },
      "post": {
        "summary": "Create an customer",
        "description": "Create an customer",
        "tags": [
          "Customer"
        ],
        "security": [
          {
            "Bearer": []
          }
        ],
        "operationId": "CreateCustomer",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "description": "Customer to create",
            "schema": {
              "$ref": "#/definitions/NewCustomerRequest"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "$ref": "#/definitions/CreateCustomerResponse"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "422": {
            "description": "Unexpected error",
            "schema": {
              "$ref": "#/definitions/GenericErrorModel"
            }
          }
        }
      }
    },
    "/customers/{customer}": {
      "get": {
        "summary": "Get detail of customer",
        "description": "Get detail customer from users",
        "tags": [
          "Customer"
        ],
        "security": [
          {
            "Bearer": []
          }
        ],
        "operationId": "GetDetailCustomer",
        "parameters": [
          {
            "name": "customer",
            "in": "path",
            "description": "Customer id of the customer to get",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/CustomerDetailResponse"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "422": {
            "description": "Unexpected error",
            "schema": {
              "$ref": "#/definitions/GenericErrorModel"
            }
          }
        }
      },
      "patch": {
        "summary": "Update an customer",
        "description": "Update an customer",
        "tags": [
          "Customer"
        ],
        "security": [
          {
            "Bearer": []
          }
        ],
        "operationId": "UpdateCustomer",
        "parameters": [
          {
            "name": "customer",
            "in": "path",
            "required": true,
            "description": "customer id of the customer to update",
            "type": "string"
          },
          {
            "name": "customers",
            "in": "body",
            "required": true,
            "description": "customer ot update",
            "schema": {
              "$ref": "#/definitions/UpdateCustomerRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/UpdateCustomerResponse"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "422": {
            "description": "Unexpected error",
            "schema": {
              "$ref": "#/definitions/GenericErrorModel"
            }
          }
        }
      },
      "delete": {
        "summary": "Delete an customer",
        "description": "Delete an customer. . Auth is required",
        "tags": [
          "Customer"
        ],
        "security": [
          {
            "Bearer": []
          }
        ],
        "operationId": "DeleteCustomer",
        "parameters": [
          {
            "name": "customer",
            "in": "path",
            "required": true,
            "description": "Customer id of the customer to delete",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/DeleteResponse"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "422": {
            "description": "Unexpected error",
            "schema": {
              "$ref": "#/definitions/GenericErrorModel"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "DeleteResponse": {
      "type": "object",
      "properties": {
        "status": {
          "type": "integer",
          "example": 200
        },
        "message": {
          "type": "string"
        }
      }
    },
...
    "CustomerDetailResponse": {
      "type": "object",
      "properties": {
        "data": {
          "type": "object",
          "properties": {
            "customer_id": {
              "type": "integer"
            },
            "name": {
              "type": "string"
            },
            "address": {
              "type": "string"
            },
            "phone_number": {
              "type": "string"
            },
            "created_at": {
              "type": "string",
              "format": "date-time"
            },
            "updated_at": {
              "type": "string",
              "format": "date-time"
            },
            "cities": {
              "type": "object",
              "properties": {
                "data": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer"
                    },
                    "name": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "UpdateCustomerRequest": {
      "properties": {
        "name": {
          "type": "string",
          "example": "Zaid Tayo"
        },
        "address": {
          "type": "string",
          "example": "JL garuda 2 no 17"
        },
        "phone_number": {
          "type": "integer",
          "example": "08213126327"
        },
        "city_id": {
          "type": "integer",
          "example": 9
        }
      },
      "required": [
        "name",
        "address",
        "phone_number",
        "city_id"
      ]
    },
    "UpdateCustomerResponse": {
      "$ref": "#/definitions/CreateCustomerResponse"
    },
    "NewCustomerRequest": {
      "properties": {
        "name": {
          "type": "string",
          "example": "Zaid"
        },
        "address": {
          "type": "string",
          "example": "JL garuda 2"
        },
        "phone_number": {
          "type": "integer",
          "example": "08123621326"
        },
        "city_id": {
          "type": "integer",
          "example": 9
        }
      },
      "required": [
        "name",
        "address",
        "phone_number",
        "city_id"
      ]
    },
    "CreateCustomerResponse": {
      "type": "object",
      "properties": {
        "success": {
          "type": "boolean"
        },
        "data": {
          "type": "object",
          "properties": {
            "name": {
              "type": "string"
            },
            "address": {
              "type": "string"
            },
            "phone_number": {
              "type": "integer"
            },
            "city_id": {
              "type": "integer"
            },
            "user_id": {
              "type": "integer"
            },
            "updated_at": {
              "type": "string",
              "format": "date-time"
            },
            "created_at": {
              "type": "string",
              "format": "date-time"
            }
          }
        },
        "message": {
          "type": "string"
        }
      }
    },
    "CustomerResponse": {
      "type": "object",
      "properties": {
        "data": {
          "$ref": "#/definitions/Customers"
        },
        "meta": {
          "$ref": "#/definitions/Pagination"
        }
      },
      "required": [
        "data"
      ]
    },
    "Customers": {
      "type": "array",
      "items": {
        "type": "object",
        "properties": {
          "customer_id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "phone_number": {
            "type": "string"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          },
          "cities": {
            "type": "object",
            "properties": {
              "data": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "integer"
                  },
                  "name": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "Pagination": {
      "type": "object",
      "properties": {
        "pagination": {
          "properties": {
            "total": {
              "type": "integer"
            },
            "count": {
              "type": "integer"
            },
            "per_page": {
              "type": "integer"
            },
            "current_page": {
              "type": "integer"
            },
            "total_pages": {
              "type": "integer"
            },
            "links": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        }
      }
    },
...
    "GenericErrorModel": {
      "type": "object",
      "properties": {
        "key": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      },
      "required": [
        "key"
      ]
    }
  }
}